PagesEditor: rename context to .js. Delete unused store.
