--- app/pages/lab-pages-editor/store.jsx
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-
-const defaultWorkflowData = {};
-const WorkflowContext = useContext(defaultWorkflowData);
-
-export default WorkflowContext;
